N70 windows support (#96)
* make Ruby optional

* fix defines for WIN

* correct EXPORT macro for Windows

* add /EHsc for windows to ffig libs

* set exception handling globally

* get rid of MSVC warning in tests

* update catch submodule and fix compile error

* require Python 3 for windows

* fix check for Python 3 on windows

* fix check for Python 3 on windows

+++ tests/src/TestShapeMocks.cpp
@@ -49,7 +49,7 @@ TEST_CASE("MockShape", "[mocks::MockShape]")
   GIVEN("A mock shape with function objects returning values")
   {
     mocks::MockShape shape;
-    size_t area_f_count = 0;
+    int area_f_count = 0;
     shape.area_ = [&area_f_count]{ ++area_f_count; return 0.0;};
 
     THEN("All method invocations return expected values through function invocations")
